PH-269: Extract `akeneo:connectivity-audit:update-data` logic into a dedicated service
