+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Audit/Command/UpdateAuditDataCommand.php
@@ -9,6 +9,7 @@ use Akeneo\Connectivity\Connection\Application\Audit\Command\UpdateDataSourcePro
 use Akeneo\Connectivity\Connection\Domain\Audit\Persistence\PurgeAuditProductQueryInterface;
 use Akeneo\Connectivity\Connection\Domain\ValueObject\HourlyInterval;
 use Akeneo\Connectivity\Connection\Infrastructure\Audit\Persistence\DbalSelectHourlyIntervalsToRefreshQuery;
+use Akeneo\Connectivity\Connection\Infrastructure\Audit\UpdateAuditData;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -20,72 +21,16 @@ use Symfony\Component\Console\Output\OutputInterface;
  */
 class UpdateAuditDataCommand extends Command
 {
-    /**
-     * @var string
-     */
     protected static $defaultName = 'akeneo:connectivity-audit:update-data';
 
-    public function __construct(
-        private UpdateDataSourceProductEventCountHandler $updateDataSourceProductEventCountHandler,
-        private DbalSelectHourlyIntervalsToRefreshQuery $selectHourlyIntervalsToRefreshQuery,
-        private PurgeAuditProductQueryInterface $purgeQuery,
-    ) {
+    public function __construct(private UpdateAuditData $updateAuditData) {
         parent::__construct();
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $this->purgeOlderThanXDays(10);
+        $this->updateAuditData->execute();
 
-        // Create a Command for the previous hour.
-        $previousHourlyInterval = HourlyInterval::createFromDateTime(
-            new \DateTimeImmutable('now -1 hour', new \DateTimeZone('UTC'))
-        );
-        $this->updateProductEventCount($previousHourlyInterval);
-
-        // Create a Command for the previous hour.
-        $previousHourlyInterval = HourlyInterval::createFromDateTime(
-            new \DateTimeImmutable('now -1 hour', new \DateTimeZone('UTC'))
-        );
-        $this->updateProductEventCount($previousHourlyInterval);
-
-        // Create a Command for the current hour.
-        $currentHourlyInterval = HourlyInterval::createFromDateTime(
-            new \DateTimeImmutable('now', new \DateTimeZone('UTC'))
-        );
-        $this->updateProductEventCount($currentHourlyInterval);
-
-        /*
-         * Create a Command for each hour retrieved from events that are not yet complete.
-         * I.e., the last update happened before the end of the event and need to be updated again.
-         */
-        $hourlyIntervalsToRefresh = $this->selectHourlyIntervalsToRefreshQuery->execute();
-        foreach ($hourlyIntervalsToRefresh as $hourlyInterval) {
-            // Ignore the current and previous hour; already added.
-            if (true === $currentHourlyInterval->equals($hourlyInterval)
-                || true === $previousHourlyInterval->equals($hourlyInterval)
-            ) {
-                continue;
-            }
-
-            $this->updateProductEventCount($hourlyInterval);
-        }
-
-        return 0;
-    }
-
-    private function purgeOlderThanXDays(int $days): void
-    {
-        $before = new \DateTimeImmutable("now - $days days", new \DateTimeZone('UTC'));
-        $before = $before->setTime((int) $before->format('H'), 0, 0);
-
-        $this->purgeQuery->execute($before);
-    }
-
-    private function updateProductEventCount(HourlyInterval $hourlyInterval): void
-    {
-        $this->updateDataSourceProductEventCountHandler->handle(
-            new UpdateDataSourceProductEventCountCommand($hourlyInterval)
-        );
+        return Command::SUCCESS;
     }
 }
